+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -9,7 +9,7 @@ import kotlin.String
  * This code is generated from build.gradle.kts
  */
 public object Versions {
-  public const val skot: String = "1.4.2-ua"
+  public const val skot: String = "1.4.3-ua"
 
   public const val group: String = "com.github.useradgents.skot"